+++ packages/signals_core/lib/signals_core.dart
@@ -3,6 +3,7 @@ export 'src/utils/timer.dart';
 export 'src/value/value.dart';
 export 'src/core/container.dart';
 export 'src/core/signals.dart';
+export 'package:preact_signals/preact_signals.dart' show SignalEffectException;
 
 // Async
 export 'src/async/connect.dart';