+++ packages/signals_core/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 6.2.0
+
+- Adding SignalsAutoDisposeMixin
+- Computed/Signal now use SignalsAutoDisposeMixin but ReadonlySignalMixin no longer has dispose methods
+- Tests now check for SignalEffectException and not error
+- Exposing SignalEffectException from preact_signals
+- Bump preact_signals to 1.9.3
+- Computed not also can throw SignalEffectException instead of generic error
+
 ## 6.1.0
 
 - Fix error in effect not showing StackTrace